refactor(text): transform text ui's to goog modules
