+++ src/os/ui/text/tuieditorlang.js
@@ -1,12 +1,12 @@
-goog.provide('os.ui.text.TuiEditorLang');
-
+goog.module('os.ui.text.TuiEditorLang');
+goog.module.declareLegacyNamespace();
 
 /**
  * Since tui editor is lazy loaded, run after loading
  * No easy way to change button text. Opened issue #524 on github
  * @param {boolean=} opt_supportHotkeys
  */
-os.ui.text.TuiEditorLang.setup = function(opt_supportHotkeys) {
+const setup = function(opt_supportHotkeys) {
   // This is the language block for tui editor so we can modify words in their UI
   const languageOverrides = {
     'Markdown': 'Text',
@@ -26,4 +26,8 @@ os.ui.text.TuiEditorLang.setup = function(opt_supportHotkeys) {
   }
 
   toastui.Editor.setLanguage('en-US', languageOverrides);
+};
+
+exports = {
+  setup
 };
